finished refactoring front end file structure
