+++ vite-project/src/Modules/Auth/components/LogoutButton.tsx
@@ -1,5 +1,5 @@
 import { useAuth0 } from "@auth0/auth0-react";
-import Button from "./Button";
+import Button from "../../../components/Button";
 
 const LogoutButton = ({ ...rest }) => {
   const { logout } = useAuth0();
